fix: WALLET-1381 — keep the SDK off the onboarding entry statically
`npm run start:chrome` failed the AssertSdkFreePageEntries check: the guard runs
in every mode, but development has no `usedExports`, so the tree-shaking it
relied on is off and onboarding really did link casper-js-sdk through
form-validation-rules.ts -> is-valid-secret-key-hash.ts. (The reported issuer,
sdk-methods.ts, is misleading — moduleGraph.getIssuer() is global.)

Move useTorusSecretKeyRule, the sole consumer of isValidSecretKeyHash, out of
form-validation-rules.ts — which is on onboarding's eager startup path — into
its own module. Onboarding is now SDK-free in the module graph itself rather
than only after production tree-shaking.
